license update and cleanup
This commit changes the dual LGPL and BSD licensing to a single
BSD-3-Clause license. This is in agreement with the dual-licensing
approach that permits choosing either license. This change is made to
simplify the somewhat confusing dual licensing.

In addition, headers are cleaned up to use SPDX based license
identifiers.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>